Add labwork5
The tasks here are implementing classes satisfy the hierarchy in which
Cylinder inherits Circle inherits Point inherits Shape.

Task 1+2 make use of abstract class.
Task 3 utilizes interface.

Note: Task 3 is subject to changed in the future, as the current model
does not apply correctly (Point shouldn't be in the same inheritance
structure with Shape).